+++ src/main/java/org/openwebnet4j/OpenDeviceType.java
@@ -67,7 +67,9 @@ public enum OpenDeviceType {
     // energy management (not defined by BTicino)
     SCS_ENERGY_METER(1830),
     // dry contacts and IR interfaces (not defined by BTicino)
-    SCS_DRY_CONTACT_IR(2510);
+    SCS_DRY_CONTACT_IR(2510),
+    // alarm (not defined by BTicino)
+    SCS_ALARM_CENTRAL_UNIT(5100);
 
     private final Integer value;
     private static Map<Integer, OpenDeviceType> mapping;